requirements.txt
absl-py==1.0.0
astunparse==1.6.3
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
cachetools==4.2.4
camel-tools==1.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
cycler==0.11.0
Cython==0.29.23
debugpy @ file:///D:/bld/debugpy_1636043436560/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1631346842025/work
dill==0.3.4
docopt==0.6.2
editdistance==0.6.0
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
filelock==3.4.0
Flask==2.0.2
flatbuffers==2.0
fonttools==4.28.3
future==0.18.2
gast==0.4.0
gensim==4.1.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.42.0
h5py==3.6.0
huggingface-hub==0.2.1
idna==3.3
importlib-metadata==4.8.2
ipykernel @ file:///D:/bld/ipykernel_1638555715117/work/dist/ipykernel-6.6.0-py3-none-any.whl
ipython @ file:///D:/bld/ipython_1638470606441/work
itsdangerous==2.0.1
jedi @ file:///D:/bld/jedi_1637175252639/work
Jinja2==3.0.3
joblib==1.1.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1637611911738/work
jupyter-core @ file:///D:/bld/jupyter_core_1636814453100/work
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1638419302549/work
nltk==3.6.5
numpy==1.21.4
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.4
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==8.4.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1639065841292/work
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1629119114968/work
pyparsing==3.0.6
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2021.3
pywin32==302
PyYAML==6.0
pyzmq @ file:///D:/bld/pyzmq_1635877415276/work
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
sacremoses==0.0.46
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
setuptools-scm==6.3.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smart-open==5.2.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.7.0
tensorflow-gpu @ https://storage.googleapis.com/tensorflow/windows/gpu/tensorflow_gpu-2.7.0-cp39-cp39-win_amd64.whl
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
threadpoolctl==3.0.0
tokenizers==0.10.3
tomli==1.2.2
torch==1.10.0
tornado @ file:///D:/bld/tornado_1635819745510/work
tqdm==4.62.3
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1635260543454/work
transformers==4.13.0
typing_extensions==4.0.1
urllib3==1.26.7
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
Werkzeug==2.0.2
wrapt==1.13.3
zipp==3.6.0
